Niobe-Standalone/pyproject.toml
[project]
name = "niobe-digital-audit"
version = "0.1.0"
description = "Lean Digital Audit platform for remote due diligence"
requires-python = ">=3.11"
dependencies = [
  "cairosvg>=2.7.1",
  "fastapi>=0.115.0",
  "httpx>=0.27.0",
  "jinja2>=3.1.4",
  "markdown>=3.6",
  "matplotlib>=3.9.0",
  "openpyxl>=3.1.5",
  "passlib[bcrypt]>=1.7.4",
  "psycopg[binary]>=3.2.1",
  "pyotp>=2.9.0",
  "pydantic>=2.8.2",
  "python-docx>=1.1.2",
  "python-multipart>=0.0.9",
  "pyyaml>=6.0.2",
  "reportlab>=4.2.2",
  "rich>=13.8.0",
  "sqlalchemy>=2.0.35",
  'tree-sitter-language-pack>=1.4.1; platform_system == "Darwin"',
  'tree-sitter-languages>=1.10.2; platform_system != "Darwin"',
  "typer>=0.12.5",
  "uvicorn>=0.30.6",
  "weasyprint>=66.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.3",
  "ruff>=0.6.4",
]

[project.scripts]
digital-audit = "Python3.cli:main"
niobe-api = "Python3.control_plane:serve"

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["Python3*"]

[tool.setuptools.package-data]
"*" = []


[tool.pytest.ini_options]
minversion = "8.0"
testpaths = ["tests"]
pythonpath = ["."]
